TVG-125 Move dropdown into ui_components folder
